requirements.txt
PyGObject>=3.42.0
pycairo>=1.20.0
rich>=13.0.0
click>=8.0.0
pyyaml>=6.0
psutil>=5.9.0

